moved 1st component into TS. failed to load class
