+++ frontend/package.json
@@ -9,8 +9,10 @@
     "@babel/core": "7.14.0",
     "@material-ui/core": "4.11.4",
     "@material-ui/icons": "4.11.2",
+    "@mui/types": "^7.1.0",
     "@stripe/react-stripe-js": "^1.6.0",
     "@stripe/stripe-js": "^1.20.3",
+    "@types/bootstrap": "^5.1.6",
     "@types/jest": "^27.0.2",
     "@types/node": "^16.11.7",
     "@types/react": "^17.0.34",
@@ -30,7 +32,8 @@
     "react-scripts": "4.0.3",
     "react-slick": "0.28.1",
     "redux": "^4.1.1",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.3.0",
+    "typescript": "^4.2.4"
   },
   "devDependencies": {
     "@babel/cli": "7.13.16",
